change global net_send_channels to uint8_t

+++ src/globals.c
@@ -7,4 +7,4 @@ int (*action)(int *argc, char *argv[]);
 int hex;
 char *progname;
 size_t blocksize = 4096;
-size_t net_send_channels = 7;
+uint8_t net_send_channels = 7;
